Return most relevent prefix search results.
Relevency is calculated in much the same way as lsearchd - based on the
number of incoming links.

This solution will also create some knock-on work, namely that we'll
have to reindex pages when the number of links to them changes
substantially.  Exactly what that means is up for debate but without
doing something a page won't notice it has become popular until it has
changed.

Bug: 52886
Change-Id: I099849fa1d9b5cd4bedef2266ac8f7ec3e31e990