Highlighting and search operator fixes.
We shouldn't be indexing html in the headings so I've stripped it.

Highlighting should be sorted by score and we need to fragment the array
fields or else get all of the ones that match back and just pick the
first.  Fragmenting them lets them get sorted by score even if they are
small.

The default search operator needs to be and.  It was and, but this was
lost in the great gerrit catastrophe of September 5, 2013.

Bug: 53955

Change-Id: I5aaf621002bef78db15c44d71a19aa84759df7f2