Include wordCount and byteSize in result
Including the wordCount and byteSize prevents the default implementation
from taking over.  The default implementation calls initText which
CirrusSearch has overridden to parse the page which is slow.

We start updating elasticsearch with a text byte size and word count and
we use those in the results if they are available, otherwise we just use
the now deprecated textLen field which is a number of bytes.

We also set some fields to stored so they should be faster to fetch.

Bug: 55592
Change-Id: I469029ef86cc629b815021103c40a878ec9758fa