Push bulk indexing onto the job queue
Add an extra parameter to forceSearchIndex.php which, when set, will
cause updates to be pushed onto the job queue and processed there.

Because it came up in code review, this renames
CirrusSearchUpdater::updateRevisions to updatePages which is more
correct - it doesn't deal in pages.  That causes a small raft of
renames.

During the rename I alsmo removed some revision related stuff in
forceSearchIndex.php which wasn't required.

Change-Id: Id17b807f9b209a7da9caf7e314e41f6dc3ff5879