Allow users prefer articles with recent changes
This can be engaged by default on the entire wiki (like for wikinews)
or left off by default.  When off by default it can be turned on by
prefixing the query with "prefer-recent:".  This engages wikinews-like
behavior on any wiki.  When on by default it can be turned off by prefixing
the query with "prefer-recent:0".

It can be further customized it by specifying the portion of the score
that decays with time like this: "prefer-recent:.1". Even more
customization can be had by specifying the half life of the decayed
portion of the score in days like this "prefer-recent:,150" or
"prefer-recent:,.1".  Both can be specified like this:
"prefer-recent:1,12".

This only effects full text search, not prefix search.

The performance cost of this is marginal.  Tested on beta's copy of
simplewiki there was more variable between repeated invocations of the
same query than between invocations of the query with and without the extra
math required for prefer-recent.  It was a few milliseconds either way.

Bug: 56690

Change-Id: I53967ce5d210c63963a3d377450c394c6373a5bd