Fix two more query types that cause ES to choke
Queries made of just spaces:  This one is probably new but didn't have a
regression test until this commit.
Queries with an escaped term in quotes (mostly "/"):  This one isn't new
and someone has been trying this query over and over and over again for
the past few days.  This should make it work.

Some small logging improvements so we can see queries that contain trailing
spaces.  Those should be trimmed but logging them will provide more
information if this breaks.

Bug: 56790
Change-Id: Ida6815a0a205a52375e3c3c7004d026d4b1a569a