Always fork Cirrus LinksUpdate behavior
Cirrus used to try to be fancy and only fork the links update behavior
when running in a web process.  This was cool but ran into trouble when
the standard links update behavior ran long causing the Cirrus behavior
to run even longer and get cut off by the maximum amount of time that
jobs are allowed to run.

This should give the Cirrus behavior the maximum possible time.  If it
still gets cut off then we have some trouble with Cirrus.

Bug: 56783
Change-Id: I1df947293a2807dcfe141d98aae73a9a016d100e