Implement per-namespace cross wiki searching
Searching works across wikis with joined namespaces.  By default results
from other wikis will be squashed from the result set unless the namespace
is NS_FILE and wfFindFile finds a file for it.  In other words, at this
point the features really is only good for NS_FILE, which is OK because
that is what we need right now.

Updates and deletes on one wiki will add or remove the marker used to filter
cross wiki duplicates correctly.

forceSearchIndex.php works for this as well - it'll always enqueue jobs
to maintain the external index and will never do it in process.

Bug: 53531
Change-Id: I42d944c769f65c746dcea74e05978ab119128311