Switch search methods to always return status
Switch search methods in CirrusSearchSearcher from returning status on
error to always return a status.  The old behaviour was compatible with
SearchEngine->searchTest but not with the prefix search hook.  It also
wasn't in line with the best practice for using status: always return it
and let the caller sort out whether it has to unwrap it.

This should stop funky warning from coming from ApiOpenSearch on Cirrus
errors.

Bug:  57961
Change-Id: I54fd705f842fcae5dfda9fe3a2c07b2bf20071c0