Stop 'store'ing fields
It turns out that you rarely want to 'store' a field in Elasticsearch
because it will load the field from the _source which is stored by default
when you want to highlighting or return something.  'store'ing a field
adds an extra disk seak to load the field when it is returned or
highlighted.  Supposedly you should 'store' the field if you plan on
frequently returning just that field but in our case, we don't even do
that for titles on prefix searches as we have to return the namespace too.

I ran some quick tests and this didn't seem to effect prefix search
performance but sped up full text search slightly.  Since it saves space
as well it sounds like a good thing to do.

Bug:  58098
Change-Id: I49d2203464ab5e8b4e9000426d2751d58b4760d6