Respect revision deletion in CirrusSearch
So with the way revision deletion works and the way that CirrusSearch
works, we're much much less likely to end up with hidden revisions in
our index--you can't suppress current revisions, so we should've already
updated the index when the page was edited.

However, Stuff Happens, and deleted revisions absolutely should never be
in the index. So we'll force a reindex on revision deletion Just In Case.

Bug: 58366
Change-Id: Ic0165595119f8ebd9e992562c4b2658b2a518706