Boost page score by templates contained on page
The boost can either be done via a configuration "message" or directly
in the query with boost-templates:"".  All tests use the query form and
the configuration documentation suggests testing the configuration using
the query form before making changes.

While there is no bug for this new feature, the implementation solves
Bug: 57112
because doing so made the implementation simpler.

Change-Id: Ib10da1d87ca5fae95f134237073f7c61eb96bed9