Smarter defaults for forceSearchIndex's batch-size
If --skipParse is specified then the batch size defaults to 500 because
larger batches are much more efficient for this and that still finishes
quite quickly.  If --skipParse is not specified then the batch size
defaults to 10 which is _hopefully_ a small enough number of pages to
fully parse in 60 seconds.

Change-Id: Ifa09fa57acd211146d39361f11c56cf643f151a6