Shrink the LinksUpdateJob a bit
This'll protect from updates that add or remove a _ton_ of links.

This moves the link selection from after the job starts to before which
means fewer links are saved to the job.  It also stores the links as
prefixed db keys rather than Titles which should make the job itself much
smaller.

Also we can afford to do about 50 links per job and still finish in a
few seconds so change the maximum number of links and unlinks to index
to 25 each.

Bug: 59123

Change-Id: Id0586b798dda57fb69350307e9d1f85b34400a7f