Combine highlighting from plain and text analyzer
For fields where this is possible, combine the highlighting into a single
snippet.  This fixes
Bug: 54526
for the most important fields: title, text, and file_text.  headings and
redirect titles will still lose their plain matches if there are stemmed
matches.  That needs Elasticsearch issue 3757 to be fixed.  The code is
there to fix that one too, but it has to go through the release process.

Change-Id: If5f10f12160f73c7d39e8607c1fe7cad4bb126bb