Don't ever ever search for nothing
So long as the user actually adds a string to the search we'll search.  Even
if their string just ends up adding a filter.  That is cool.  We just
don't want to search when that box is empty.  That takes a long time because
we score most documents in the index and then throw them away because the
user was just looking to load the page.

Bug: 59821

Change-Id: I54b0cefe6ec052ade39c56440455bea274b06090