Remove log10 from boost in prefix search
Because nothing but number of incoming links effects the score removing
the log10 from boost sorts the results the same way while doing less work.

On itwiki this amounts for going from 200ms for a prefix search for "a"
to 60ms.  Even though these are cached most of the time this is worth it.

Change-Id: Ifc60f466f324161b2349aecbe1312aef7993d172