Return timestamp from index rather than database
This gives us a really quick way to the last revision that made it into
Cirrus.

Fun fact: the only reason we now load the latest revision of all results
is to check if there is one in the \SearchResult->isMissingRevision.  I
wonder if we can improve page load times by skipping even that somehow....

Bug: 60004

Change-Id: I2c8917167ce4e49ba820661eedcc610d636ce67d