Handle near match conflicts
This handles the case where Elasticsearch returns more than one "near match"
by taking somewhat educated guesses as to the "right choice" and if no tactic
yields a single title that is better then the other then it declares that
there is no near match and takes you to the search screen.

Steps:
1.  Only one match?  That one.
2.  Just one string match perfectly?  That one.
3.  Just one string match but for case?  That one.
4.  Just one string match if you Title Case all titles and the query?  That
one.
5.  Give up and send the user to the search page.

Bug: 59841

Change-Id: I1a6039fca51574bd3f46bb701e15816898e3c3db