Never run the plain highlighter on fuzzy queries
Fuzzy queries are super duper slow on the plain highlighter so just skip
using it for those queries.  This means that queries that use ~ to get
a fuzzy match won't have highlighted headings or redirects but they'll be
much faster and less likely to anger Elasticsearch's CPUs.

In my development environment these queries are ~3700% faster which seems
worth it for missing out on the less used highlighted fields.

Change-Id: I96b1dd736c57ca02c5f98572d553f242cef7a047