Always search plain text and stemmed text
This has several advantanges:
1.  Finds both exact matches and stemmed matches but sorts the exact
matches higher.  Yay.
2.  Finds articles that only contain stop words!  Sweet!
3.  Highlights stopwords in results.  Very nice!

It has two problems, though:
1. Stop words will be worth as much as exact matching text.  They really
ought to be worth less than stemmed text.  Like 10% or something.  Anyway
that requires some work in Elasticsearch to get that hopping.  I mean, stop
words in the article text will still be worth very little because they
are common but they will be uncommon in things like headings and titles
which will make them worth more when they appear.
2. Stop words are now required.  They used to be ignored which was bad
but now they are required which, I think, is less bad but still not good.
The side effect here is that if you search for "the once and future king"
you won't find an article named "once and future king" if the article
doesn't contain the word "the".  This is somewhat moot because more
articles (in English) will likely contain the word "the".  On enwiki, the
article for "The Once and Future King" even contains the word "The" in the
title.

BUG: 60302
BUG: 54937
Change-Id: Ic1d1e494832435308135392b5774614f24a710b8