Support SearchEngine's new sort parameter
Sorting by a field requires that the field be loaded into memory so I
want to be careful not to let all fields be sortable but the title
should be safe because it is small compared to, say, text.

Won't do anything without Ie946150c6796139201221dfa6f7750c210e97166 in core
but won't crash either.

Change-Id: I4708391aed8325cf3f082b7de7e438efa031d2cb