Browser test gardening
* Split large features into smaller ones.  This is nice because parallel
cucumbers runs each feature in parallel.
* Add a screenshooter that snaps a shot on failure if you set an environment
variable.
* Update test README.

The tests now all pass on PhantomJS.  I'm pretty sure they already did but
splitting them up made this useful.  PhantomJS is a bit flakey with parallel
cucumber but it is worth it for the speed up.  I'm finding that it works
well to run everything in parallel with PhantonJS:
 export BROWSER_LABEL=phantomjs
 bundle exec parallel_cucumber -n5 features
then gather the failures and rerun those in sequence:
 bundle exec cucumber features/blah.feature:10 features/boo.feature:193

Change-Id: Ib84bb19c96ae37fa4c666b1227044264ab3b5c4f