Catch parse errors from Elasticsearch and retry
This is a stopgap that prevents users from getting the "temporary error"
message nearly as frequently.  We'll make some effort to run their query
even though we didn't fully understand it.  It is still an error to send
something to Elasticsearch that causes it to through back a parse error
but at least this has a chance helping the user when we have an error.

Bug: 56877
Change-Id: I00c3fd920bd8a27921beac58bd8817986badef01