Tweak search config for performance
This turns off leading wildards (? and *) and weakens fuzzy queryies (~)
to only match terms that share the first two characters with the term
marked as fuzzy.  Both of these should speed up queries by bounding the
portion of the terms that Lucene will have to check.  Without this we
scan across a huge portion of the terms dictionary with both of these
which is slow.

Modified wildcard tests to run for both ? and * and covered this case.

Change-Id: Ic6a84da7ddbccbb1756f8bea84bf013494ef950d