Combine all incoming link counts into one field
Combine all incoming link counts into incoming_links.

Turn off 'index'ing for some of the long fields that we only want to store
and retrieve or use for boosting.

These two save a tiny amount of space.  They'll save more when we can
drop the incoming_redirect_links field entirely but not too much.  The
real advantage of this is that we will be able to switch prefix search to a
sort from a script based boost over a constant score query.

Change-Id: I8894858a5d349a93b2dfe109188a13eb2881840d