Turn on create and suggestions query has syntax
Detect syntax like incategory: and cata?ult.  If the query contains any
of that then don't offer the user a link to make a page for the search.
If the query contains any of the latter kind of syntax then don't try to
suggest alternate spellings to the user because the suggestions aren't
likely to be very good.

Bug: 60630
Change-Id: I873ab914cca0fb60710fae82d0c84a78a818c1aa