Use near_match to also search pages
This helps us find pages with text that Elasticsearch isn't handling properly
in another way.  Like pages named ¢.

Note that near_matches count twice as much as regular matches which should
boost "perfect" matches pretty effectively.

Note also that we don't make an effort to highlight the near_match matches.
Partly this is because they'll likely already be highlighted by the regular
highlighter.  The problem with getting the rest is that we'd need to use
the FVH for them which requires additional storage which we don't want to
spend at this point.  For the most part, this is an improvement.

Bug: 61080
Change-Id: I631efa8d426b75766f7a5b149e3ee41048e5e5a0