Handle _s as spaces in prefix:
We've already fixed underscores vs spaces in namespaces but this will fix
it for pages.  The reason this slipped through for so long is that core
handles the underscore and space munging for prefix search but doesn't
get involved with the prefix: syntax which was built to replicate it.
Things slipped through the cracks.

Bug: 30125
Change-Id: I9c7be71190633f693659c455d8dbd8855df93067