Fix some odd query issues caused by near_matches
Using the near_match analyzer was causing some pages not to be found when
we expected they should be.  The regression tests caught it.  I must not
have run them all before merging.

In any case, this fixes the problem by splitting the near match from the
main query_string and dropping it into a bool query.  This makes a stricter
"or" between the two things: near_match matching and full text matching.

Change-Id: Ib324ecfaf2e80379d54669b25e3f73be0ca4c4ca