Squash master to REL1_22
Preserves mediawiki 1.22 compatibility.

Backport support for extracting file text.

Disable tests for not showing suggestions if the page is found or the search
contains syntax.  That requires 1.23 to work.

Keep workaround for lack of apibeforemain hook so we still provide prefix
search results.

A few of the integration tests still fail but they are mostly related to
1.23 core changes and aren't indicators of large problems.

Note: this includes cirrus' efforts to replace titlekey.  You shuold be
able to turn of titlekey once you have this version of cirrus.

Change-Id: I67eeadfc929099077f60d86dbfc2702ca9f95972