Hardcode interwiki searches offset and limit
Trying to take these from the parent was insane and very likely
to be a performance drain on larger limits with many wikis as
interwiki sources.

Instead just always provide the first 10 at most. We're only trying
to give people a glimpse of the sister projects to encourage users
to click through.

Change-Id: I7984cd60d44eecbab6d818429809c76df068593d