Only suggest words that share two first characters
In my testing for work on speeding up the phrase suggester in Elaticsearch
I found that adding a larger prefix drasticly improves suggestion performance
at the cost of not considering errors in the first two characters rather than
just not considering errors in the first.  5x performance increase is not
unlikely.

This will generate fewer suggestions but that _should_ be worth it.

Change-Id: Iab246c8a1a0a784ef98a3063b9db015687bf58b0