More job reattempts, this time LinksUpdateJob(Prioritized)
updatePages() returns a count of the updated pages. We can cleverly
use this to tell if our single page updated (batch is harder why I
haven't fixed MassIndexJob yet)

Return false (and early) when we failed to update so we can reattempt.
This way we'll also delay LinksUpdateSecondary until after we've
made sure we've finished the primary job.

Change-Id: I5c3e80f8f37cc5e7dab7560ad5616be71726d99b