Add support for the experimental highlighter
This highlighter must be installed as an Elasticsearch plugin before you
can use it.  Once you have you can turn it on and use it without making
any changes to the index.  You can also instruct Cirrus to optimize the
index for the experimental highlighter.  If you do the next time you reindex
highlights will speed up and the index should shrink.  If you turn off
the experimental highlighter now that you've optimized the index for it
searches will crash.  To turn it off you'll have to turn off the optimization,
reindex, and then turn it off.

Got 54411 in the process because it was in the way.

Bug: 60141
Bug: 54411
Bug: 54526

Change-Id: Ie546c1b50e6394b8f100766d3fb7ee1b5a0aaf8e