send_some improvements
Fix an off by one error that would cause it to replay at twice real time.

Switch from an "every nth log" to a "p% of logs" configuration.  This is
more granular when you really want to replay 40%, 50%, 70% or whatever.

Also make it possible to replay the search against the wiki it was against
in the first place.

Change-Id: I090fcd7e552497f6381989ff237f51e05b50503c