Add an auxiliary text field
This field will hold text that isn't part of the article proper but still
should be searched.  Stuff like caption titles and see more links.  It'll
be searched with lower boost then page text and highlighted only if there
isn't any page text to highlight.

DEPLOYMENT: This requires rebuilding the index.  If you don't rebuild the
index portions of the page text will be written to the auxiliary_text field
but won't be searchable or highlightable.  This shouldn't be big problem
because we expect the text to be less commonly searched.  We'll rebuild the
indexes quickly anyway.

NOTE:  We're relying on field autocreation being off for the above statement
to be true.  That is OK because it is off at WMF and the README tells folks
to turn it off.  If it isn't off then the first update will autocreate the
field and it won't be highlightable.  In fact, it'll cause queries to fail.
All of them.  Don't do that.

REQUIRES: I97cf66014719244b8bb2b0509b419c82202bdb01 from core

Change-Id: I70be621d2702409d88befd8f7ba0936cbc4912ef