Boost results that contain hits in the opening
This is done by indexing the opening section into another field and querying
it.  We don't need to highlight it because we already highlight the article
text _and_ we already boost early hits in the text in the highlighter.  The
highlighter uses fixed windows but this works ok.

DEPLOYMENT: Requires adding the new field and doing a from mediawiki reindex.
The new field should be safe to add by running updateSearchIndexConfig.php
without the reindex parameters.

Bug: 61669
Bug: 66045
Change-Id: Ib7ec7a114f5b2d063df90c10c1ea5afb8a49af3e