Search categories
This should provide a nice boost for pages that are categorized with matching
terms.  It doesn't add highlighting for categories.  We're filing Bug 66426 to
do that.  We're not doing it now because of the performance hit for
highlighting yet another field.  For now if a hit comes only from the
category then the user will get a snippet from the front of the article.

We also remove the code to filter out categories from the page text.  The
categories aren't in the page text anyway.

Deployment: Deploying this will break incategory until the wiki is in place
reindexed.  This should hopefully be ok.

Bug: 66045
Change-Id: I9157c9e7cb4d42bd18be5fd176eed2b4ec822e2a