Don't update the index if it hasn't changed
If the document doesn't change then don't push a new copy of it into the
index.  Elasticsearch doesn't do this by default and updates are very
expensive.  This should prevent any that aren't really required.  I have an
instinct that says this is a pretty significant fraction of updates at WMF.
Instinct only, though.  We graph it in ganglia so we can be sure if it is
worth it.

Requires https://github.com/ruflin/Elastica/pull/629 or we'll lose retry
on conflict support.

Change-Id: I533e08ac15d5e48d1de8b36e3b86a419f339be38