Insource support
This supports pre-tokenized search and regex search.

Caveats:
* The regex syntax is
http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/query-dsl-regexp-query.html
* The regex search is pretty slow.  We're using a somewhat nasty hack to
speed it up quite a bit but it'll never be as fast as pre-tokenized search.
* Doesn't support highlighting.  I've filed a bug with the experimental
highlighter to try to get regex highlighting support.  I don't imagine
that'll actually be too difficult.
* The regex is case sensitive by default.  It can be made case insensitive
using insource://i but that has higher overhead.  Not sure if the overhead
is too high or not a big deal given the rest of the overhead of regexes.
Either way, its an option now that we can try.

DEPLOYMENT: Regex searches will duck out of the pool counter unless you
add a new pool counter config called CirrusSearch-Regex.  These queries
are much heavier then normal searches so best would be to allow only one
or two concurrent regex searches with a backlog of 10 or 20 or something.
The low backlog helps prevent some abuse by clogging apaches waiting on
a slow search.

Bug: 65783
Change-Id: I9745b828700edaa416fe38a8f19d7987f52a7b57