Properly handle pages that turn into redirects
When a page goes from being normal to a redirect without being moved we
used to leave it in the index.  It'd show up in search results and when
people clicked on it they'd be taken to the target of the redirect.  The
snippet would come from whatever was in the page right before it was turned
into a redirect.  Lame.

This removes the redirect from the index by a delete.  We aren't careful
to make sure that the redirect is in the index to be deleted because that
is more complex and marginally slower.  So far as I can tell there are few
advantages to this "speculative" delete behavior we're using.

Bug: 61211
Change-Id: Iee168e2719127d655a4d211edd6bd02caca566cf