Add a maintenance script to make the index sane
The idea is to quickly look at every page and check for insanity:
* Page exists but isn't in index
* Page is rediret but is in index
* Page doesn't exist and is in index
* Page is in more then one index

Once it detects insanity it (mostly) just queues a job to to fix the index
for the page.

Also adds the wiki name to forceSearchIndex.php.

Requires Iee168e2719127d655a4d211edd6bd02caca566cf or won't work from
redirects in the index.

Bug: 61211
Bug: 66243
Change-Id: Iafbdbfbde5bf631ec394a0af973e0948d45ea436