Fix some Special:Random bugs
1.  If a user asked for a random file in a namespace (Special:Random/File
for example), then we'd use the namespace to pick an index but not as a
filter.  This caused searches in the main namespace to work ok but not
those in, say, File.  They'd get results from User_talk and other crazy
stuff.
2.  Searches in the File: namespace were picking up results from commons.
While we want searches to pick up results from commons we probably don't
want Special:Random to do it.

Bug: 66643

Change-Id: I4a8e6eb0979584866ff3b9aa3aeddf646544c779