Fix logging for non-existent updates
If trying to update a document that doesn't exist that's not really a fatal
error.  We try to detect it and log it at the debug level.

We imagine this could happen if you add a link to the page right after
creating it and the update count job makes its way there before the page
creation job does.  In that case we're happy to just skip the link count
update job because it'll be covered by the page creation job.

This also happens during the second stage of bootstrapping where you just
index the links.  If there was an error during the first stage and the
page itself wasn't indexed or, maybe, the page is newly created and not yet
in the index.

Change-Id: Ifa4fe2e86ec57a5276e9d3dc4df453e6d17155ad