Configure cache warmers
By default we configure a cache warmer in the content index for the main
page because we know that'll find _something_.  Beyond that we allow extra
warmers to be configured by adding them to an array.  The main page warmer
can also be disabled if it isn't useful.

Also adds an option to updateSearchIndexConfig to support _just_ adding
the warmer without checking the rest of the config.  This is useful if you
want to be super sure that you will only update the warmers without change
the index contents.  Its also quick.

Bug: 67094

Change-Id: I379303310e1877159a2f965c4973df05fde12c50