Use combined all field for near_match
This is less important then the combined all fields for plain and text
matching, but it is still helpful.

Also, fix some issues with fuzzy queries using the all fields and with
the opening weighting using the all fields.  It was 8 which made opening
text 10 (8 for opening + 2 for text) which was too high compared to headings.
Lowered it to 6.

Bug: 68167
Change-Id: Id561b44e6461bb4719a9ee42350d80619fbdbc85