Eat bad forward slashes
We're not sure why, but forward slashes aren't always escaped correctly.
Sometimes they turn into regex queries which can get pretty slow.  It seems
to be environment specific - development doesn't do it.  Might be hhvm or
the version of php-curl or ubuntu or whatever.  We don't know.  Rather then
try to send the / escaped its simpler to just turn it into a space.

Change-Id: Icb6a2f0097dc25f354dabae3632cfe962a21341b