Add configurable delay to our most frequent jobs
The idea is that we can use the delay as a knob to cut down on updates
from repeated edits or constant template changes.  This works because while
a job is delayed it still counts for deduplication - another copy of the
job will not make it onto the queue.  Instead the first copy remains delayed
until its timeout and then it runs.

I'm not actually sure that we'll want to engage this in production but it
gives us the option and can serve as a stepping stone for something better.

Change-Id: I77fa71297c5e26e3cf4b38d4c8fce1072f5443c2