Shrink the weighted all field's positions
We need to store position information for the weighted all fields and
right now we store position information for the entire article text twice
and that is big!  This has us store it once but at a cost: now file text
and aux text is worth the same as article text.  We still highlight it
later which is good.

The disk size reduction is pretty good but where this _should_ really
shine is in using less disk IO.  The less disk IO should be worth the
loss of precision.

Change-Id: I84612b829b00aa4790d9c53ce44dc944ea67ea1f