Fix total_shards_per_node config
Before this was only being applied after the index was created.  Now it
gets a special value if the index is being reindexed to spread the index
as evenly as possible during the write traffic reindex event.  Once the
reindex is done the value is set to whatever the user configured it to
be.

Change-Id: I9807023baa2fa8c4e4a5a79cf38d1c372462ddfb