Redo job that counts incoming links
Old way:
When a page gets a link added or removed queue a single job with the target
of those links in it.  That job counts all the links to each of the updated
pages and reindexes all of them.

New way:
When a page gets a link added or removed queue one job per link target that
was added or removed.  That job counts all the links for only a single
page.

The new way is better then the old way because it allows deduplication of
links count update jobs.  This is important if a popular template is edited
to add or remove links to a certain page.  In the old way the incoming links
were counted over and over and over again and the page might be updated
many time though many will be nooped because the incoming links count won't
have changed enough.  In the new way the page will be updated and the links
counted far fewer times due to deduplication.

Change-Id: Ie1ddf414f0bcb188dfa536ea9050aeff5f410cf5