Add support for ngram accelerated regexes
Its disabled by default because it requires the wikimedia/extra elasticsearch
plugin and casual installers won't have it.  It must be enabled in two stages:
1.  build
2.  use

In the build stage when the index is rebuilt the source text is analyzed
with an additional analyzer that is required for regex acceleration.  It
does not require the wikimedia/extra plugin to be installed on Elasticsearch.

In the use stage the newly analyzed field is used to accelerate regular
expressions.  If the field isn't yet built all regular expressions will
return no return no results.  If the wikimedia/extra plugin isn't installed
in Elasticsearch this will cause all regular expression searches to fail.

Change-Id: If9fcb6a03c23e8abaf59b5ddfadab051263dac33
DEPLOYMENT: Lots of notes.  Read commit message.